more changes to pass CI (missing module import, removed manipulate impl)
