+++ src/new_module/selection.jl
@@ -333,39 +333,4 @@ function normalize_selection(idx::AbstractIndex,
     return combine_src => fun => newcol
 end
 
-function manipulate(df::Any, @nospecialize(cs...); copycols::Bool, keeprows::Bool, renamecols::Bool)
-    cs_vec = []
-    for v in cs
-        if v isa AbstractVecOrMat{<:Pair}
-            append!(cs_vec, v)
-        else
-            push!(cs_vec, v)
-        end
-    end
-    normalized_cs = Any[normalize_selection(index(df), make_pair_concrete(c), renamecols) for c in cs_vec]
-    return _manipulate(df, normalized_cs, copycols, keeprows)
-end
-
-_manipulate(df, normalized_cs::Vector{Any}, copycols::Bool, keeprows::Bool) = throw("struct needs to define this")
-
-
-function manipulate(df, args::AbstractVector{Int};
-                    copycols::Bool, keeprows::Bool, renamecols::Bool)
-    throw("struct needs to implement this")
-end
-
-function manipulate(df, c::MultiColumnIndex; copycols::Bool, keeprows::Bool,
-                    renamecols::Bool)
-    if c isa AbstractVector{<:Pair}
-        return manipulate(df, c..., copycols=copycols, keeprows=keeprows,
-                          renamecols=renamecols)
-    else
-        return manipulate(df, index(df)[c], copycols=copycols, keeprows=keeprows,
-                          renamecols=renamecols)
-    end
-end
-
-
-manipulate(df, c::ColumnIndex; copycols::Bool, keeprows::Bool,
-           renamecols::Bool) =
-    manipulate(df, Int[index(df)[c]], copycols=copycols, keeprows=keeprows, renamecols=renamecols)
+function manipulate end
