+++ src/new_module_index/SeparateModuleIndex.jl
@@ -1,4 +1,6 @@
 module SeparateModuleIndex
+using Unicode: Unicode
+using REPL: REPL
 using DataAPI: All, Between, Cols, DataAPI, nrow
 using InvertedIndices: InvertedIndices, Not
 using Printf: @sprintf
@@ -7,11 +9,11 @@ include("utils.jl")
 include("index.jl")
 
 # new def here; the above files are copied as is (stuff removed from utils.jl)
-index(df) = throw("struct needs to define this")
+function index end
 
 export
     _findall, add_names, funname, SubIndex, _names,
     AsTable, ColumnIndex, MultiColumnIndex, Index,
     COLUMNINDEX_STR, MULTICOLUMNINDEX_STR, SymbolOrString, AbstractIndex,
-    MULTICOLUMNINDEX_TUPLE, make_unique, index,parentcols,rename!,fuzzymatch
+    MULTICOLUMNINDEX_TUPLE, make_unique, index,parentcols,rename!,fuzzymatch, make_unique!, _norm_eq, _julia_charmap
 end
