+++ test/runtests.jl
@@ -14,12 +14,12 @@ else
     @show Threads.nthreads()
 end
 
-# ambiguities_vec = Test.detect_ambiguities(DataFrames, recursive=true)
-# if !isempty(ambiguities_vec)
-#     @error "Method ambiguities:"
-#     display(ambiguities_vec)
-#     throw(AssertionError("method dispatch ambiguities found"))
-# end
+ambiguities_vec = Test.detect_ambiguities(DataFrames, recursive=true)
+if !isempty(ambiguities_vec)
+    @error "Method ambiguities:"
+    display(ambiguities_vec)
+    throw(AssertionError("method dispatch ambiguities found"))
+end
 
 unbound_args_vec = Test.detect_unbound_args(DataFrames, recursive=true)
 if !isempty(unbound_args_vec)
